requirements.txt
absl-py==0.7.1
aioredis==1.2.0
appnope==0.1.0
asgiref==3.1.2
asn1crypto==0.24.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
backcall==0.1.0
bleach==1.5.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.3.3
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
Django==2.2.2
django-background-task==0.1.8
django-background-tasks==1.2.0
django-compat==1.0.15
django-extensions==2.1.9
django-ide==0.0.5
gast==0.2.2
grpcio==1.21.1
guesslang==0.9.3
hiredis==1.0.0
html5lib==0.9999999
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1.1
msgpack==0.6.1
numpy==1.16.4
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.6.0
tensorflow==1.7.0rc1
termcolor==1.1.0
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
wcwidth==0.1.7
Werkzeug==0.15.4
zope.interface==4.6.0
